requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
mysqlclient==2.0.3
protobuf==3.17.3
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
